[MIG/REF] subcontracted_service: Migration to 14.0
