+++ subcontracted_service/models/__init__.py
@@ -1,3 +1,3 @@
 from . import product
-from . import procurement
+from . import procurement_group
 from . import warehouse
