+++ subcontracted_service/readme/CONFIGURE.rst
@@ -1,9 +1,8 @@
 To configure this module, you need to:
 
-#. Configure your service product with the flag
-   ``property_subcontracted_service`` in product form if this product should
-   trigger a procurement.
-#. Add supplier in your product form.
+#. Configure your service product with the flag *Subcontracted Service*
+   in the product form if this product should trigger a procurement.
+#. Add the supplier information in your product form in the *Purchase* tab.
 #. Additionally and despite a predefined rule is created in each warehouse,
    you can configure the 'Subcontracting_service procurement rule' for each
    warehouse through 'Inventory / Configuration / Warehouse Management /
