+++ subcontracted_service/readme/CONTRIBUTORS.rst
@@ -1,4 +1,4 @@
 * Damien Crier <damien.crier@camptocamp.com>
-* Jordi Ballester Alomar <jordi.ballester@eficent.com>
-* Lois Rilo <lois.rilo@eficent.com>
+* Jordi Ballester Alomar <jordi.ballester@forgeflow.com>
+* Lois Rilo <lois.rilo@forgeflow.com>
 * Rubén Bravo <rubenred18@gmail.com>
